// libs/client_sdk/models/create-prompt-version-response-dto.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */

// May contain unused imports in some cases
// @ts-ignore
import type { PromptVersionDto } from './prompt-version-dto';

export interface CreatePromptVersionResponseDto {
  promptId: string;
  version: PromptVersionDto;
  versionCount: number;
}